fix a few compiler warnings (thanks Martin)
git-svn-id: https://serveur-svn.lri.fr/svn/modhel/luatex/trunk@3064 0b2b3880-5936-4365-a048-eb17d2e5a6bf

+++ source/texk/web2c/luatexdir/pdf/pdfgen.c
@@ -1618,7 +1618,7 @@ void check_o_mode(PDF pdf, char *s, int o_mode_bitpattern, boolean strict)
             assert(0);
         }
         snprintf(warn_string, 99, "not allowed in %s mode (\\pdfpoutput = %d)",
-                 m, pdf_output);
+                 m, (int)pdf_output);
         if (strict)
             pdf_error(s, warn_string);
         else
